Make generator dependencies an explicit option
Adding a directory as a "Files" item would make changes to that
directory trigger a regeneration of the project. Not only did was
this fairly hidden behaviour, it also required stating all files
in the emission phase to check if they were directories or not
which added a decent chunk to total generation time.
(At the scanning phase files likely are stated anyway when
globing folders so no need to do it twice.)