Massive source restructure.
The single build.h has been split into multiple headers grouped
in folders to get a more logical structure of the code and
dependencies for readability. There still is a build.h that
includes everything in a single header fashion, but it's now
also possible to some extent to only include parts as needed.