Make the file writer only write if data is different
When outputting data it's important to not touch modification
time of files if nothing changed. The simple file writer now
compares the incoming data with existing data on disk (if any)
and only writes the file if it's different. While it's always
faster to just not generate the data if not needed, this way is
often much simpler and safer.